refactor(ml-1162): split IAT integration tests into own file
Move the integration describe block out of controller.test.js into a
new controller.integration.test.js. The unit test (pure handler +
h.view mock) stays in controller.test.js. The integration file boots
the real server via setupTestServer and asserts on rendered HTML.

This split will scale better as more IAT pages land under
journey/self-service/.
